+++ include/dolfin/math/LinearDistribution.h
@@ -10,74 +10,78 @@
 namespace dolfin
 {
 
+//-----------------------------------------------------------------------------
+
 struct LinearDistribution
 {
 
-  uint global_size{0};
-  uint card{0};
-  uint rank{0};
-  uint L{0};
-  uint R{0};
-  uint offset{0};
-  uint size{0};
+  size_t global_size { 0 };
+  size_t card { 0 };
+  size_t rank { 0 };
+  size_t L { 0 };
+  size_t R { 0 };
+  size_t offset { 0 };
+  size_t size { 0 };
 
   ///
   LinearDistribution() = default;
 
   ///
-  LinearDistribution(uint global_size, uint card, uint rank):
-    global_size(0),
-    card(0),
-    rank(0),
-    L(0),
-    R(0),
-    offset(0),
-    size(0)
+  LinearDistribution( size_t global_size, size_t card, size_t rank )
+    : global_size( 0 )
+    , card( 0 )
+    , rank( 0 )
+    , L( 0 )
+    , R( 0 )
+    , offset( 0 )
+    , size( 0 )
   {
-    set(global_size, card, rank);
+    set( global_size, card, rank );
   }
 
   ///
-  inline bool in_range(uint index)
+  inline bool in_range( size_t index )
   {
-    return ((offset <= index) && (index < offset + size));
+    return ( ( offset <= index ) && ( index < offset + size ) );
   }
 
   ///
-  inline uint owner(uint index) const
+  inline size_t owner( size_t index ) const
   {
-    return static_cast<uint>(std::max(
-        std::floor((real) index / (real) (L + 1)),
-        std::floor((real) ((real) index - (real) R) / (real) L)));
+    return static_cast< size_t >( std::max(
+      std::floor( ( real ) index / ( real )( L + 1 ) ),
+      std::floor( ( real )( ( real ) index - ( real ) R ) / ( real ) L ) ) );
   }
 
   ///
-  void set(uint global_size, uint card, uint rank)
+  void set( size_t global_size, size_t card, size_t rank )
   {
-    this->global_size = global_size,
-    this->card    = card;
-    this->rank    = rank;
-    this->L       = std::floor((real) global_size / (real) card);
-    this->R       = global_size % card;
-    this->offset  = rank * L + std::min(rank,R);
-    this->size    = std::floor(((real) global_size + (real) card - (real) rank - 1.0) / (real) card);
+    this->global_size = global_size, this->card = card;
+    this->rank   = rank;
+    this->L      = std::floor( ( real ) global_size / ( real ) card );
+    this->R      = global_size % card;
+    this->offset = rank * L + std::min( rank, R );
+    this->size =
+      std::floor( ( ( real ) global_size + ( real ) card - ( real ) rank - 1.0 )
+                  / ( real ) card );
   }
 
   ///
   void disp() const
   {
-    section("LinearDistribution");
-    message("global size    : %u", global_size);
-    message("num partitions : %u", card);
-    message("rank           : %u", rank);
-    message("offset         : %u", offset);
-    message("size           : %u", size);
-    message("quotient size  : %u", L);
-    message("remainder      : %u", R);
+    section( "LinearDistribution" );
+    message( "global size    : %u", global_size );
+    message( "num partitions : %u", card );
+    message( "rank           : %u", rank );
+    message( "offset         : %u", offset );
+    message( "size           : %u", size );
+    message( "quotient size  : %u", L );
+    message( "remainder      : %u", R );
     end();
   }
-
 };
+
+//-----------------------------------------------------------------------------
 
 } /* namespace dolfin */
